refactor: Modularize LLM tools and clean up config
This commit refactors LLM tools into a dedicated `app/llm_tools` package, separating them into `course_tools.py` and `misc_tools.py` for better organization.

- Renamed `list_full_courses` to `list_courses_and_details` for clarity.
- Removed unused variables from `config.py`.
- Updated the user timezone in the system instructions to "America/Recife".
