+++ server/app/llm_tools/__init__.py
@@ -1,0 +1,6 @@
+from .course_tools import course_tools
+from .misc_tools import misc_tools
+
+tools = []
+tools.extend(course_tools)
+tools.extend(misc_tools)
