+++ server/app/services/__init__.py
@@ -1,2 +1,1 @@
 from .google_ai_service import get_google_ai_service, GoogleAIService
-from .llm_tools import tools
